requirements.txt
langchain==0.2.14
langchain-chroma==0.1.2
langchain-community==0.2.12
setuptools==73.0.1
streamlit==1.37.1
python-dotenv==1.0.1